+++ programs/turf_vault/src/instructions/create_user_account.rs
@@ -19,7 +19,11 @@ pub struct CreateUserAccount<'info> {
     pub system_program: Program<'info, System>,
 }
 
-pub fn handle_create_user_account(ctx: Context<CreateUserAccount>, wallet: Pubkey) -> Result<()> {
+pub fn handle_create_user_account(
+    ctx: Context<CreateUserAccount>,
+    wallet: Pubkey,
+    username: [u8; 32],
+) -> Result<()> {
     let user = &mut ctx.accounts.user_account;
     user.wallet = wallet;
     user.balance = 0;
@@ -27,6 +31,7 @@ pub fn handle_create_user_account(ctx: Context<CreateUserAccount>, wallet: Pubke
     user.total_withdrawn = 0;
     user.total_won = 0;
     user.seeds = 0;
+    user.username = username;
     user.bump = ctx.bumps.user_account;
 
     msg!("User account created for: {}", wallet);